Feat: 기술, 직무 도메인을 경력 도메인 하위로 이동
