+++ src/main/kotlin/com/interrupt/server/career/job/repository/JobGroupRepository.kt
@@ -1,0 +1,9 @@
+package com.interrupt.server.career.job.repository
+
+import com.interrupt.server.career.job.entity.JobGroup
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface JobGroupRepository: JpaRepository<JobGroup, Long> {
+}
