+++ src/main/kotlin/com/interrupt/server/career/skill/dto/SkillGroupDto.kt
@@ -1,0 +1,16 @@
+package com.interrupt.server.career.skill.dto
+
+import com.interrupt.server.career.skill.entity.SkillGroup
+import com.interrupt.server.common.annotation.RedisEntity
+
+@RedisEntity
+data class SkillGroupDto(
+    val id: Long,
+    val name: String,
+) {
+
+    val skillList: MutableList<SkillDto> = mutableListOf()
+
+    fun toEntity(): SkillGroup = SkillGroup(name)
+
+}
