+++ src/main/kotlin/com/interrupt/server/career/skill/repository/SkillQueryRepository.kt
@@ -1,0 +1,32 @@
+package com.interrupt.server.career.skill.repository
+
+import com.interrupt.server.career.job.dto.JobGroupDto
+import com.interrupt.server.career.skill.dto.SkillDto
+import com.interrupt.server.career.skill.entity.Skill
+import com.interrupt.server.career.skill.entity.SkillGroup
+import org.springframework.stereotype.Repository
+
+@Repository
+class SkillQueryRepository(
+   private val skillRepository: SkillRepository
+) {
+
+    fun findJobDtoList(): List<SkillDto> = skillRepository.findAll {
+        select(
+            new(
+                SkillDto::class,
+                path(Skill::id),
+                path(Skill::name),
+                new(
+                    JobGroupDto::class,
+                    path(SkillGroup::id),
+                    path(SkillGroup::name),
+                )
+            )
+        ).from(
+            entity(Skill::class),
+            join(Skill::skillGroup).`as`(entity(SkillGroup::class))
+        )
+    }.filterNotNull()
+
+}
