--- src/main/kotlin/com/interrupt/server/job/entity/Job.kt
@@ -1,18 +1,0 @@
-package com.interrupt.server.job.entity
-
-import com.interrupt.server.common.entity.SoftDeleteBaseEntity
-import jakarta.persistence.*
-
-@Entity
-@Table(
-    name = "job",
-    uniqueConstraints = [UniqueConstraint(name = "uk_job_name", columnNames = ["name"])]
-)
-class Job(
-    @field:Column(name = "name", nullable = false, unique = true)
-    var name: String,
-    @field:ManyToOne(fetch = FetchType.LAZY)
-    @field:JoinColumn(name = "job_group_id", nullable = false)
-    var jobGroup: JobGroup
-): SoftDeleteBaseEntity() {
-}
