--- src/main/kotlin/com/interrupt/server/job/repository/MemberJobRepository.kt
@@ -1,9 +1,0 @@
-package com.interrupt.server.job.repository
-
-import com.interrupt.server.job.entity.MemberJob
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface MemberJobRepository: JpaRepository<MemberJob, Long> {
-}
