+++ src/main/kotlin/com/interrupt/server/member/repository/MemberSkillRepository.kt
@@ -1,6 +1,6 @@
 package com.interrupt.server.member.repository
 
-import com.interrupt.server.skill.entity.MemberSkill
+import com.interrupt.server.career.skill.entity.MemberSkill
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
