--- src/main/kotlin/com/interrupt/server/skill/dto/SkillDto.kt
@@ -1,15 +1,0 @@
-package com.interrupt.server.skill.dto
-
-import com.interrupt.server.skill.entity.Skill
-import com.interrupt.server.common.annotation.RedisEntity
-
-@RedisEntity
-data class SkillDto(
-    val id: Long,
-    val name: String,
-    val skillGroup: SkillGroupDto
-) {
-
-    fun toEntity(): Skill = Skill(name, skillGroup.toEntity())
-
-}
