--- src/main/kotlin/com/interrupt/server/skill/repository/SkillGroupRepository.kt
@@ -1,9 +1,0 @@
-package com.interrupt.server.skill.repository
-
-import com.interrupt.server.skill.entity.SkillGroup
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface SkillGroupRepository: JpaRepository<SkillGroup, Long> {
-}
